+++ contrib/pageinspect/btreefuncs.c
@@ -206,14 +206,12 @@ check_relation_block_range(Relation rel, int64 blkno)
 	if (blkno < 0 || blkno > MaxBlockNumber)
 		ereport(ERROR,
 				(errcode(ERRCODE_INVALID_PARAMETER_VALUE),
-				 errmsg("invalid block number %lld",
-						(long long) blkno)));
+				 errmsg("invalid block number %" PRId64, blkno)));
 
 	if ((BlockNumber) (blkno) >= RelationGetNumberOfBlocks(rel))
 		ereport(ERROR,
 				(errcode(ERRCODE_INVALID_PARAMETER_VALUE),
-				 errmsg("block number %lld is out of range",
-						(long long) blkno)));
+				 errmsg("block number %" PRId64 " is out of range", blkno)));
 }
 
 /* -----------------------------------------------
