+++ src/backend/access/rmgrdesc/clogdesc.c
@@ -28,15 +28,15 @@ clog_desc(StringInfo buf, XLogReaderState *record)
 		int64		pageno;
 
 		memcpy(&pageno, rec, sizeof(pageno));
-		appendStringInfo(buf, "page %lld", (long long) pageno);
+		appendStringInfo(buf, "page %" PRId64, pageno);
 	}
 	else if (info == CLOG_TRUNCATE)
 	{
 		xl_clog_truncate xlrec;
 
 		memcpy(&xlrec, rec, sizeof(xl_clog_truncate));
-		appendStringInfo(buf, "page %lld; oldestXact %u",
-						 (long long) xlrec.pageno, xlrec.oldestXact);
+		appendStringInfo(buf, "page %" PRId64 "; oldestXact %u",
+						 xlrec.pageno, xlrec.oldestXact);
 	}
 }
