+++ src/backend/access/rmgrdesc/mxactdesc.c
@@ -58,7 +58,7 @@ multixact_desc(StringInfo buf, XLogReaderState *record)
 		int64		pageno;
 
 		memcpy(&pageno, rec, sizeof(pageno));
-		appendStringInfo(buf, "%lld", (long long) pageno);
+		appendStringInfo(buf, "%" PRId64, pageno);
 	}
 	else if (info == XLOG_MULTIXACT_CREATE_ID)
 	{
