+++ src/backend/replication/slot.c
@@ -1544,11 +1544,11 @@ ReportSlotInvalidation(ReplicationSlotInvalidationCause cause,
 	{
 		case RS_INVAL_WAL_REMOVED:
 			{
-				unsigned long long ex = oldestLSN - restart_lsn;
+				uint64		ex = oldestLSN - restart_lsn;
 
 				appendStringInfo(&err_detail,
-								 ngettext("The slot's restart_lsn %X/%X exceeds the limit by %llu byte.",
-										  "The slot's restart_lsn %X/%X exceeds the limit by %llu bytes.",
+								 ngettext("The slot's restart_lsn %X/%X exceeds the limit by %" PRIu64 " byte.",
+										  "The slot's restart_lsn %X/%X exceeds the limit by %" PRIu64 " bytes.",
 										  ex),
 								 LSN_FORMAT_ARGS(restart_lsn),
 								 ex);
